fix(ci): select predictive validation without PR packages
