+++ .github/workflows/linux-binaries.yml
@@ -83,6 +83,13 @@ jobs:
         uses: oven-sh/setup-bun@v2
         with:
           bun-version: "1.3.14"
+          no-cache: true
+
+      - name: Setup Node.js
+        uses: actions/setup-node@v5
+        with:
+          node-version-file: package.json
+          package-manager-cache: false
 
       - name: Setup Node.js
         uses: actions/setup-node@v5
